Sim loop created. Not tested, but each key step should be there and in the right order.

+++ solenoid-seperator-sim/sim.h
@@ -17,7 +17,7 @@
 // the main function will also initialise the objects, as an init function for each object was a waste of time
 
 // this will update each object every timestep, for the duration of the simulation
-void updateSystem(solenoid *, capactitor *, ejectionObject *, float, int);
+void updateSystem(solenoid *, capactitor *, ejectionObject *, float, float);
 
 // this will delete each object from dynamic memory
 void cleanUp(solenoid *, capactitor *, ejectionObject *);
